refactor: extract duplicated tracks data to shared module
- Create shared tracks data module at data/tracks.ts
- Remove duplicated tracks constant from TrackList.tsx and BookmarksScreen.tsx
- Update imports to use shared tracks data and Track type
