+++ components/TrackList.tsx
@@ -6,6 +6,7 @@ import { RouteProp } from '@react-navigation/native';
 import { MaterialIcons } from '@expo/vector-icons';
 import { Audio } from 'expo-av';
 import AppHeader from '@components/AppHeader';
+import { tracks, Track } from '../data/tracks';
 
 // 型定義
 export type RootStackParamList = {
@@ -14,26 +15,7 @@ export type RootStackParamList = {
   Bookmarks: { bookmarks: string[] };
 };
 
-export type Track = {
-  id: string;
-  title: string;
-  artist: string;
-  artwork: any;
-  file: any;
-  duration?: string; // 追加: 曲の長さ
-};
-
-const tracks: Track[] = [
-  {
-    id: '1',
-    title: 'Sample Track',
-    artist: 'Sample Artist',
-    artwork: require('../assets/icon.png'),
-    file: require('../assets/sample.mp3'),
-    duration: '2:34', // 追加: 固定の時間
-  },
-  // 追加のトラックはここに
-];
+export { Track } from '../data/tracks';
 
 export default function TrackList() {
   const navigation = useNavigation<NativeStackNavigationProp<RootStackParamList, 'TrackList'>>();
